+++ src/cartpole/test/test_mpc.py
@@ -1,0 +1,8 @@
+import pytest
+
+
+@pytest.mark.skip(reason="No test has been created yet")
+@pytest.mark.copyright
+@pytest.mark.linter
+def test_pid():
+    assert 1 == 2
